+++ sshd-core/src/main/java/org/apache/sshd/client/auth/pubkey/KeyPairIdentity.java
@@ -41,8 +41,8 @@ public class KeyPairIdentity implements PublicKeyIdentity {
 
     public KeyPairIdentity(SignatureFactoriesManager primary, SignatureFactoriesManager secondary, KeyPair pair) {
         this.signatureFactories = ValidateUtils.checkNotNullAndNotEmpty(
-                SignatureFactoriesManager.resolveSignatureFactories(primary, secondary),
-                "No available signature factories");
+            SignatureFactoriesManager.resolveSignatureFactories(primary, secondary),
+            "No available signature factories");
         this.pair = Objects.requireNonNull(pair, "No key pair");
     }